+++ test/docs_test.rb
@@ -113,7 +113,7 @@ class DocsTest < Minitest::Test
   end
 
   def assert_examples(method)
-    return if [:align_frames, :from_epoch, :coalesce, :cumsum, :arange, :concat, :lengths, :is_nan, :join, :is_not_nan, :read_csv_batched].include?(method.name)
+    return if [:align_frames, :from_epoch, :coalesce, :cumsum, :arange, :concat, :lengths, :is_nan, :join, :is_not_nan, :read_csv_batched, :sink_parquet].include?(method.name)
 
     # TODO fix
     return if method.name == :groupby_dynamic